More changes to classes.

+++ src/main/java/com/abcc/maven/jmath/NumberAdditions.java
@@ -2,16 +2,16 @@ package com.abcc.maven.jmath;
 
 public class NumberAdditions {
 	
-	public static long add2numbers (long iNum1, long iNum2) {
+	public static int add2numbers (int iNum1, int iNum2) {
 	
-		long	iResult = iNum1 + iNum2;
+		int	iResult = iNum1 + iNum2;
 				
 	return iResult;
 	}
 	
-	public static long add3numbers (long iNum1, long iNum2, long iNum3) {
+	public static int add3numbers (int iNum1, int iNum2, int iNum3) {
 		
-		long	iResult = iNum1 + iNum2 + iNum3;
+		int	iResult = iNum1 + iNum2 + iNum3;
 				
 	return iResult;
 	}
